fix: correct f32/f64 precision mismatches in DNN math functions
In C, calling math functions like exp(), log(), log10(), sqrt(), pow()
on float arguments implicitly promotes them to double. The Rust code
was calling f32 methods directly (e.g. f32::log10(), f32::powf()),
producing different results.

Fixed in:
- freq.rs: lpc_from_cepstrum pow(10,x), compute_burg_cepstrum pow(.995,i)
  and log10() — all now use f64 math matching C's implicit promotion
- lpcnet.rs: compute_frame_features sqrt(), exp(), ln() — promoted to f64
- fargan.rs: run_fargan_subframe exp() — promoted to f64

Also fixed Clippy warnings in osce_nndsp test.

+++ src/dnn/fargan.rs
@@ -449,7 +449,8 @@ fn run_fargan_subframe(st: &mut FARGANState, pcm: &mut [f32], cond: &[f32], peri
         cond,
         ACTIVATION_LINEAR,
     );
-    let gain = gain[0].exp();
+    // C: gain = exp(gain) — float promoted to double, result truncated to float.
+    let gain = (gain[0] as f64).exp() as f32;
     let gain_1 = 1.0 / (1e-5 + gain);
 
     // Build pitch prediction and previous samples
